app/requirements.txt
Django==3.2.6
gunicorn
django-filter>
djangorestframework
Markdown
django-cors-headers>
dj-database-url>
djangorestframework-simplejwt
django-ses
coverage
flake8
flake8-docstrings
pytest
pytest-xdist
pytest-cov
pytest-django
psycopg2
celery
ipython
model-mommy
tox
autopep8